fix(rtp): key the track maps on 32 bits and stop inserting null jitter buffers
_tracks and the two jitter buffer maps are keyed by track id, but the key type was
uint8_t. The WebRTC provider passes the SSRC as track id (the RTSP one passes the channel
id), so the key was silently truncated to its low byte: two SSRCs of the same stream
sharing that byte collide and overwrite each other's entry, breaking the track. With an
audio and a video SSRC that is roughly a 0.4% chance per session.

SetContentMediaType() also reached the map through operator[], which inserts a null
shared_ptr for an unknown key - and OnRtpReceived() only tests the iterator against
end(), so it would dereference it. It now uses find(), checks for null, and returns false
with a warning instead of silently doing nothing.

+++ src/projects/modules/rtp_rtcp/rtp_rtcp.h
@@ -89,12 +89,15 @@ private:
 	std::shared_ptr<RtcpTransportCcFeedbackGenerator> _transport_cc_generator = nullptr;
 
 	// Jitter buffer
-	// payload type : Jitter buffer
-	std::unordered_map<uint8_t, std::shared_ptr<RtpFrameJitterBuffer>> _rtp_frame_jitter_buffers;
-	std::unordered_map<uint8_t, std::shared_ptr<RtpMinimalJitterBuffer>> _rtp_minimal_jitter_buffers;
+	// track id : Jitter buffer
+	// The key must be 32 bits wide: the WebRTC provider uses the SSRC as track id (the RTSP one
+	// uses the channel id), and a uint8_t key silently truncated it - two SSRCs sharing their
+	// low byte would collide and overwrite each other's track.
+	std::unordered_map<uint32_t, std::shared_ptr<RtpFrameJitterBuffer>> _rtp_frame_jitter_buffers;
+	std::unordered_map<uint32_t, std::shared_ptr<RtpMinimalJitterBuffer>> _rtp_minimal_jitter_buffers;
 
-	// payload type : MediaTrack Info
-	std::unordered_map<uint8_t, std::shared_ptr<MediaTrack>> _tracks;
+	// track id : MediaTrack Info
+	std::unordered_map<uint32_t, std::shared_ptr<MediaTrack>> _tracks;
 	bool _video_receiver_enabled = false;
 	bool _audio_receiver_enabled = false;
